release 3.1.6 source code for java

+++ services/cdn/src/main/java/com/huaweicloud/sdk/cdn/v1/model/Urls.java
@@ -13,7 +13,7 @@ public class Urls {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "id")
 
-    private Integer id;
+    private Long id;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "url")
@@ -50,7 +50,7 @@ public class Urls {
 
     private String fileType;
 
-    public Urls withId(Integer id) {
+    public Urls withId(Long id) {
         this.id = id;
         return this;
     }
@@ -59,11 +59,11 @@ public class Urls {
      * urlid
      * @return id
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
